backend/requirements.txt
flask==3.0.2
flask-socketio==5.3.6
flask-cors==4.0.2
python-dotenv==1.0.1
python-dateutil==2.8.2
requests==2.31.0
watchdog==3.0.0
pandas==2.1.1
joblib==1.3.2
